Add regex to remove version appendix
This is just an improvised fix to get the script working with the recent 11.3 release, where appearently no appendix is used (yet?). Fetching the OS version for the jail fails due to the unknown appendix used. This fix only considered working if you run 11.3-RELEASE.
More investigation is needed to solve this issue properly.