Useful tweaks
required files - I recommend a check for the existence of the Caddy binary and Caddyfile. This is a built-in check. With the check, a service caddy start will alert the user if either of these files is missing; without it, no cue is provided.
Superfluous command definitions - stop and restart, as well as many other commands such as status and extracommands, are automatically supported as defined in rc(8). See https://www.freebsd.org/cgi/man.cgi?query=rc&sektion=8&manpath=freebsd-release-ports. There doesn't appear to be any value in including these command definitions. If anything, it makes things a little more ambiguous as it detracts from the commands that must be defined or altered i.e. start, validate and reload.